Convert time.Time object to the local time zone before sending to the MySQL server.
Before this change, storing a time.Time object with a non-local
timezone in a database and then retrieving the object again resulted
in a time.Time object denoting a different point in time.

Example: the time "2013-08-09 21:30:43 +0800 CST" was stored in a
DATETIME column of the database as "2013-08-09 21:30:43".  Reading
the value back resulted in "2013-08-09 21:30:43 +0100 BST".

This commit fixes the problem by always converting time.Time objects
to the local timezone before sending them to the MySQL server.