MYSQL_TYPE_TIMESTAMP: fix Microsecond alignment & use Microseconds instead of Nanoseconds
According the mysql docuemtation here:
http://dev.mysql.com/doc/internals/en/binary-protocol-value.html
"ProtocolBinary::MYSQL_TYPE_TIMESTAMP" expects the final 4 bytes of an 11 byte MYSQL_TYPE_TIMESTAMP to contain a little endien unsigned 32 byte value representing microseconds. Prior to this change, this value was based on nanoseconds and also misaligned (it occupied bytes 7-10 instead of bytes 8-11, with byte 7 later overwritten with seconds as represented by single byte).

This change both properly aligns the value and converts from Nanoseconds to Microseconds and back as appropriate when writing to and reading from mysql.