Add re-prepare logic to stmt Exec
Added logic to Exec, ExecFirst and ExecLast to re-prepare statements on
recovered connections. This was added as although dropped connections
are re-established, the (previously created) statements are no longer
valid. Although this could be caught by client code, the usefulness of
autorc (which is pretty much a utility) would be diminished.

Note: The re-prepare logic could be implemented by the connection
(instead of the statement) by having letting the connection keep track
of the previously successfully created statements. However, this would
greatly complicate logic on re-prepares (e.g. the connection could drop
midway through). In light of this, a lazy approach on the actual
statements was chosen. The solution could be better, but doing so would
implicate a broader refactor.